backend/requirements.txt
Django
djangorestframework
mysqlclient
python-dotenv
django-cors-headers